added payroll

+++ HRLeaveManagement/HRLeaveManagment.Application/DTOs/EmployeeDetails/UpdateEmployeeDetailsDto.cs
@@ -8,10 +8,10 @@ namespace HRLeaveManagment.Application.DTOs.EmployeeDetails
         public string Address2 { get; set; }
         public string Town { get; set; }
         public string PostCode { get; set; }
-        public int WorkPhoneNumber { get; set; }
-        public int HomePhoneNumber { get; set; }
-        public int MobileNumber { get; set; }
-        public int WorkMobileNumber { get; set; }
+        public long WorkPhoneNumber { get; set; }
+        public long HomePhoneNumber { get; set; }
+        public long MobileNumber { get; set; }
+        public long WorkMobileNumber { get; set; }
         public string PersonalEmail { get; set; }
         public string NINumber { get; set; }
         public string TaxCode { get; set; }
